Se refactorizo metodos para validar los PathVariable(required true or false) de los Controllers.

+++ src/main/java/com/udacity/jdnd/course3/critter/pet/service/PetService.java
@@ -74,7 +74,7 @@ public class PetService {
 
         return null;
     }
-    public CustomJasperReport generatePetsReport(int numberPet){
+    public CustomJasperReport generatePetsReport(long numberPet){
         List<Pet> pets = mascotas();
         if(Objects.isNull(pets))
             throw new NullPointerException();
